Refactor: Rename and refactor styles
