+++ components/svgs/Arrow.jsx
@@ -1,6 +1,6 @@
 const Arrow = ({ className }) => {
     return (
-        <svg className={`w-2.79 h-2.79 duration-500  ${className}`} width="15" height="15" viewBox="0 0 15 15" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <svg className={`w-2.79 h-2.79 duration-500 ${className}`} width="15" height="15" viewBox="0 0 15 15" fill="none" xmlns="http://www.w3.org/2000/svg">
             <path d="M2.86365 1.13672L14 1.13672V12.2731" stroke="black" strokeWidth="1.5" />
             <line x1="13.894" y1="1.03033" x2="0.530341" y2="14.394" stroke="black" strokeWidth="1.5" />
         </svg>
