+++ data/offers.js
@@ -1,7 +1,7 @@
-import polygonImage from "../public/assets/images/polygon.webp";
-import ballImage from "../public/assets/images/ball.webp";
-import arcImage from "../public/assets/images/arc.webp";
-import diamondImage from "../public/assets/images/diamond.webp";
+import polygonImage from "/public/assets/images/polygon.webp";
+import ballImage from "/public/assets/images/ball.webp";
+import arcImage from "/public/assets/images/arc.webp";
+import diamondImage from "/public/assets/images/diamond.webp";
 
 const offers = [
     {
