+++ pages/_app.js
@@ -1,7 +1,7 @@
 import "../styles/globals.css";
 import localFont from "@next/font/local";
 const neufileGrotesk = localFont({
-  src: "../public/assets/font/NeufileGrotesk-Regular.woff2",
+  src: "../public/assets/fonts/NeufileGrotesk-Regular.woff2",
 });
 export default function App({ Component, pageProps }) {
   return (
