+++ tailwind.config.js
@@ -96,7 +96,6 @@ module.exports = {
         17.63: "4.4063rem",
         18.75: "4.688rem",
         19.75: "4.9375rem",
-        20: "5rem",
         20.25: "5.0625rem",
         22.13: "5.5313rem",
         23: "5.75rem",
@@ -107,7 +106,6 @@ module.exports = {
         27.5: "6.875rem",
         30.75: "7.688rem",
         32.75: "8.1875rem",
-        40: "10rem",
         43.68: "10.91873rem",
         44.75: "11.1875rem",
         47.25: "11.8125rem",
@@ -128,7 +126,7 @@ module.exports = {
         107.5: "26.875rem",
         115.75: "28.9375rem",
         141.25: "35.3125rem",
-        147.5: "39.375rem",
+        147.5: "36.875rem",
         150: "37.5rem",
         153.5: "38.375rem",
         157.5: "39.375rem",
@@ -139,9 +137,6 @@ module.exports = {
         239: "59.75rem",
         295.25: "73.8125rem",
         320: "80rem",
-        74.69: "74.69%",
-        87.38: "87.38%",
-        87.65: "87.65%",
         "3.73/5": "74.69%",
         "4.37/5": "87.38%",
         "4.38/5": "87.65%",
@@ -150,8 +145,8 @@ module.exports = {
       maxWidth: {
         97.5: "24.375rem",
         107.5: "26.875rem",
-        147.5: "39.375rem",
-        70: "24.375rem",
+        // 157.5: "39.375rem",
+        // 70: "17.5rem",
         233: "58.25rem",
         320: "80rem",
       },
